Research Data/Experimental/Experimental-2954648d4d35e4014a65fee1737259e7e5ef525c/nlp/tools/langchain-file-processor/requirements.txt
streamlit
langchain
langchain-chroma
langchain_openai
python-dotenv
chromadb
pypdf
